Renamed help to information; moved helpcommand to new command folder
