+++ src/main/java/me/partlysanestudios/partlysaneskies/features/information/WikiArticleOpener.java
@@ -3,7 +3,7 @@
 // See LICENSE for copyright and license notices.
 //
 
-package me.partlysanestudios.partlysaneskies.features.help;
+package me.partlysanestudios.partlysaneskies.features.information;
 
 import me.partlysanestudios.partlysaneskies.PartlySaneSkies;
 import me.partlysanestudios.partlysaneskies.features.economy.auctionhousemenu.AuctionHouseGui;
